Update README.md
Water tank overflow is a common problem which leads to the wastage of water. Though there are many solutions to it like ball valves which automatically stop the water flow once the tank gets full. but being a tech. enthusiastic wouldn't you like an electronic solution for it? so here is a simple and handy project that will guide you to make a circuit which will detect the water level and will indicate the level upon getting the water tank full or a present level water level indicator is a modern way of measuring the water level using latest technologies like sensors, Arduino. The main aim of the project is to calculate the water level at any instant of time and let the pump on once, the water level is very low or the tank is almost empty. I have used Arduino UNO microcontroller to make it possible, this may be useful to conserve water, and helps us not to waste it. The user can monitor the level of water in the tank and the status of the pump being ON or OFF on Blynk -IoT App installed on their Android phones.